Structure code for example codes
