+++ components/BasicExampleProjects/kanban/files/index.html.ts
@@ -1,0 +1,15 @@
+export const language = 'markup';
+export const content = `<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" href="/favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>vuejs-kanban-yorkie</title>
+  </head>
+
+  <body>
+    <div id="app" class="kanban"></div>
+    <script type="module" src="/src/main.js"></script>
+  </body>
+</html>`;
