+++ components/BasicExampleProjects/kanban/files/main.js.ts
@@ -1,0 +1,7 @@
+export const language = 'javascript';
+export const content = `import { createApp } from 'vue'
+import App from './App.vue'
+
+import './assets/main.css'
+
+createApp(App).mount('#app')`;
