+++ components/BasicExampleProjects/kanban/files/package.json.ts
@@ -1,0 +1,19 @@
+export const language = 'json';
+export const content = `{
+    "name": "vuejs-kanban-yorkie",
+    "version": "0.0.0",
+    "homepage": "https://hunkim98.github.io/vuejs-kanban-yorkie",
+    "scripts": {
+        "dev": "vite",
+        "build": "vite build",
+        "preview": "vite preview"
+    },
+    "dependencies": {
+        "vue": "^3.2.41",
+        "yorkie-js-sdk": "^0.2.19"
+    },
+    "devDependencies": {
+        "@vitejs/plugin-vue": "^3.1.2",
+        "vite": "^3.1.8"
+    }
+}`;
