+++ components/BasicExampleProjects/kanban/index.ts
@@ -1,0 +1,2 @@
+export * from './documentStructure';
+export * from './projectCode';
